修改删除群组

+++ TimeBack/src/main/java/com/manager/master/service/serviceImpl/GroupServicelmpl.java
@@ -56,6 +56,11 @@ public class GroupServicelmpl implements IGroupService {
     private RemindJpaRepository remindJpaRepository;
 
 
+    /**
+     * 参与人列表查询
+     * @param inDto
+     * @return
+     */
     @Override
     public List<GroupOutDto> selectAll(GroupFindInDto inDto) {
         int userId = inDto.getUserId();
@@ -253,7 +258,11 @@ public class GroupServicelmpl implements IGroupService {
         }
         List<GtdGroupMemberEntity> groupMemberEntities=null;
         if(findType==2) {
+            //获取该群组下所有群成员
             groupMemberEntities = groupMemberRepository.findMemberByGroupId(groupId);
+            if (groupMemberEntities == null || groupMemberEntities.size() == 0 ) {
+                throw new ServiceException("该群组群成员数据为空");
+            }
         }
 
         List<GroupMemberOutDto> list=new ArrayList<>();
@@ -325,8 +334,7 @@ public class GroupServicelmpl implements IGroupService {
         if (groupHeadImgUrl == null || "".equals(groupHeadImgUrl)) throw new ServiceException("群头像不能为空");
         if (member.size() == 0 || member == null) throw new ServiceException("群员不能为空");
         Date date = new Date();
-        List<GtdGroupMemberEntity> groupMembers = new ArrayList<>();
-        List<GtdGroupLabel> groupLabels = new ArrayList<>();
+
 
         for(Integer i:labelId){
             if(labelId.size()!=1&&i== FIND_GROUP_LABELTYPE){
@@ -435,9 +443,17 @@ public class GroupServicelmpl implements IGroupService {
         GtdGroupEntity group = null;
         try {
             group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
+            if (group == null) {
+                throw new ServiceException("该群组数据为空");
+            }
+
         } catch (Exception e) {
             throw new ServiceException("群组不存在");
         }
+
+        //根据群组ID获取群组下所有用户ID
+        List<Integer> userIds=groupMemberRepository.findAllUserIdByGroupId(groupId);
+
         int createId=group.getCreateId();
         if(userId==createId) {//判断是否为群组创建人
             boolean flag = false; //判断是否为权限群组
@@ -450,30 +466,34 @@ public class GroupServicelmpl implements IGroupService {
             if (flag) {//TODO 权限群组发送通知 删除群组
                 //根据群组ID获得该群组的日程
                 List<GtdGroupScheduleEntity> groupSchedules=groupScheduleJpaRepository.findGroupSchedulesByGroupId(groupId);
-                for(GtdGroupScheduleEntity g:groupSchedules){
-                    int scheduleId=g.getScheduleId();//获取日程ID
+                if(groupSchedules.size()!=0&&groupSchedules!=null) {
+                    for (GtdGroupScheduleEntity g : groupSchedules) {
+                        int scheduleId = g.getScheduleId();//获取日程ID
 
-                    boolean boo=true;
-                    //根据日程ID获取有该日程的所有群组 判断这些群组中有没有此用户
-                    List<GtdGroupScheduleEntity> groups=groupScheduleJpaRepository.findGroupSchedulesByGroupId(scheduleId);
-                    for(GtdGroupScheduleEntity gs:groups){
-                        if(gs.getGroupId()!=groupId){
-                            List<GtdGroupMemberEntity> groupMembers=groupMemberRepository.findMemberByGroupId(gs.getGroupId());
-                            for(GtdGroupMemberEntity member:groupMembers){
-                                if(member.getUserId()==userId){
-                                    boo=false;
+                        boolean boo = true;
+                        //根据日程ID获取有该日程的所有群组 判断这些群组中有没有此用户
+                        List<GtdGroupScheduleEntity> groups = groupScheduleJpaRepository.findGroupSchedulesByGroupId(scheduleId);
+                        for (GtdGroupScheduleEntity gs : groups) {
+                            if (gs.getGroupId() != groupId) {
+                                List<GtdGroupMemberEntity> groupMembers = groupMemberRepository.findMemberByGroupId(gs.getGroupId());
+                                for (GtdGroupMemberEntity member : groupMembers) {
+                                    if (member.getUserId() == userId) {
+                                        boo = false;
+                                    }
                                 }
                             }
                         }
-                    }
 
-                    //在其他此日程群组不存在就删除
-                    if(boo) {
-                        List<Integer> playerIds = schedulePlayersJpaRepository.findAllPlayersId(scheduleId);
-                        for(Integer i:playerIds) {
-                            remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                        //在其他此日程群组不存在就删除
+                        if (boo) {
+                            List<Integer> playerIds = schedulePlayersJpaRepository.findAllPlayersId(scheduleId);
+                            for (Integer i : playerIds) {
+                                remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                            }
+                            for(Integer i:userIds) {
+                                schedulePlayersJpaRepository.deleteConnectionByScheduleIdAndUserId(scheduleId,i);//删除参与人表
+                            }
                         }
-                        schedulePlayersJpaRepository.deleteConnectionByScheduleId(scheduleId);//删除参与人表
                     }
                 }
                groupRepository.deleteByGroupId(groupId);//删除群组、群组标签、群成员、群组日程
@@ -502,9 +522,15 @@ public class GroupServicelmpl implements IGroupService {
         GtdGroupEntity group = null;
         try {
             group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
+            if (group == null) {
+                throw new ServiceException("该群组数据为空");
+            }
         } catch (Exception e) {
             throw new ServiceException("群组不存在");
         }
+        //根据群组ID获取群组下所有用户ID
+        List<Integer> userIds=groupMemberRepository.findAllUserIdByGroupId(groupId);
+
         int createId=group.getCreateId();
         if(userId!=createId) {
             //不是群组创建人
@@ -512,10 +538,12 @@ public class GroupServicelmpl implements IGroupService {
             for(GtdGroupScheduleEntity g:groupSchedules) {
                 int scheduleId = g.getScheduleId();
                 List<Integer> playerIds = schedulePlayersJpaRepository.findAllPlayersId(scheduleId);
-                for (Integer i : playerIds) {
-                    remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                if(playerIds.size()!=0&&playerIds!=null) {
+                    for (Integer i : playerIds) {
+                        remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                    }
+                    schedulePlayersJpaRepository.deleteConnectionByScheduleIdAndUserId(scheduleId,userId);//删除参与人表
                 }
-                schedulePlayersJpaRepository.deleteConnectionByScheduleId(scheduleId);//删除参与人表
             }
         }else{
             //删除群组
@@ -529,30 +557,34 @@ public class GroupServicelmpl implements IGroupService {
             if (flag) {//TODO 权限群组发送通知 删除群组
                 //根据群组ID获得该群组的日程
                 List<GtdGroupScheduleEntity> groupSchedules=groupScheduleJpaRepository.findGroupSchedulesByGroupId(groupId);
-                for(GtdGroupScheduleEntity g:groupSchedules){
-                    int scheduleId=g.getScheduleId();//获取日程ID
+                if(groupSchedules!=null&&groupSchedules.size()!=0) {
+                    for (GtdGroupScheduleEntity g : groupSchedules) {
+                        int scheduleId = g.getScheduleId();//获取日程ID
 
-                    boolean boo=true;
-                    //根据日程ID获取有该日程的所有群组 判断这些群组中有没有此用户
-                    List<GtdGroupScheduleEntity> groups=groupScheduleJpaRepository.findGroupSchedulesByGroupId(scheduleId);
-                    for(GtdGroupScheduleEntity gs:groups){
-                        if(gs.getGroupId()!=groupId){
-                            List<GtdGroupMemberEntity> groupMembers=groupMemberRepository.findMemberByGroupId(gs.getGroupId());
-                            for(GtdGroupMemberEntity member:groupMembers){
-                                if(member.getUserId()==userId){
-                                    boo=false;
+                        boolean boo = true;
+                        //根据日程ID获取有该日程的所有群组 判断这些群组中有没有此用户
+                        List<GtdGroupScheduleEntity> groups = groupScheduleJpaRepository.findGroupSchedulesByGroupId(scheduleId);
+                        for (GtdGroupScheduleEntity gs : groups) {
+                            if (gs.getGroupId() != groupId) {
+                                List<GtdGroupMemberEntity> groupMembers = groupMemberRepository.findMemberByGroupId(gs.getGroupId());
+                                for (GtdGroupMemberEntity member : groupMembers) {
+                                    if (member.getUserId() == userId) {
+                                        boo = false;
+                                    }
                                 }
                             }
                         }
-                    }
 
-                    //在其他此日程群组不存在就删除
-                    if(boo) {
-                        List<Integer> playerIds = schedulePlayersJpaRepository.findAllPlayersId(scheduleId);
-                        for(Integer i:playerIds) {
-                            remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                        //在其他此日程群组不存在就删除
+                        if (boo) {
+                            List<Integer> playerIds = schedulePlayersJpaRepository.findAllPlayersId(scheduleId);
+                            for (Integer i : playerIds) {
+                                remindJpaRepository.deleteAllByPlayersId(i);//删除提醒时间表
+                            }
+                            for(Integer i:userIds) {
+                                schedulePlayersJpaRepository.deleteConnectionByScheduleIdAndUserId(scheduleId,i);//删除参与人表
+                            }
                         }
-                        schedulePlayersJpaRepository.deleteConnectionByScheduleId(scheduleId);//删除参与人表
                     }
                 }
                 groupRepository.deleteByGroupId(groupId);//删除群组、群组标签、群成员、群组日程
@@ -585,6 +617,9 @@ public class GroupServicelmpl implements IGroupService {
         if (groupName == null || "".equals(groupName)) throw new ServiceException("群组名不能为空");
         if (groupHeadImgUrl == null || "".equals(groupHeadImgUrl)) throw new ServiceException("群头像不能为空");
         GtdGroupEntity group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
+        if (group == null) {
+            throw new ServiceException("该群组数据为空");
+        }
         //GtdGroupLabel groupLabel=
         int createId = group.getCreateId();
         Set<GtdLabelEntity> labels = group.getLabel();//群组原标签
@@ -631,10 +666,12 @@ public class GroupServicelmpl implements IGroupService {
                     // TODO 变为权限群组 给群员发通知 组成员状态默认为未接受(2)
                     groupJpaRepository.save(group);
                     List<GtdGroupMemberEntity> list = groupMemberRepository.findMemberByGroupId(groupId);
-                    for (GtdGroupMemberEntity g : list) {
-                        g.setGroupMemberStatus(2);
+                    if(list.size()!=0&&list!=null) {
+                        for (GtdGroupMemberEntity g : list) {
+                            g.setGroupMemberStatus(2);
+                        }
+                        groupMemberRepository.saveAll(list);
                     }
-                    groupMemberRepository.saveAll(list);
                 }
             } else {//权限群
                 if (!status) {//不加权限标签
@@ -665,7 +702,9 @@ public class GroupServicelmpl implements IGroupService {
         if (groupId == 0 || "".equals(groupId)) throw new ServiceException("群组ID不能为空");
         if (member.size() == 0 || member == null) throw new ServiceException("群员不能为空");
         GtdGroupEntity group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
-
+        if (group == null) {
+            throw new ServiceException("该群组数据为空");
+        }
         boolean flag = false; //判断是否为权限群组
         Set<GtdLabelEntity> labels = group.getLabel();
         for (GtdLabelEntity label : labels) {
@@ -730,6 +769,9 @@ public class GroupServicelmpl implements IGroupService {
         if (member.size() == 0 || member == null) throw new ServiceException("群员不能为空");
 
         GtdGroupEntity group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
+        if (group == null ) {
+            throw new ServiceException("该群组数据为空");
+        }
         boolean flag = false; //判断是否为权限群组
         Set<GtdLabelEntity> labels = group.getLabel();
         for (GtdLabelEntity label : labels) {
@@ -740,19 +782,20 @@ public class GroupServicelmpl implements IGroupService {
 
         if (!flag) {
             for (GroupMemberDto gmDto : member) {
-                if (gmDto.getUserId() == 0 || "".equals(gmDto.getUserId())) throw new ServiceException("群员ID不能为空");
+               // if (gmDto.getUserId() == 0 || "".equals(gmDto.getUserId())) throw new ServiceException("群员ID不能为空");
                 if (gmDto.getUserContact() == null || "".equals(gmDto.getUserContact()))
                     throw new ServiceException("群员联系方式不能为空");
                 if (gmDto.getUserName() == null || "".equals(gmDto.getUserName()))
                     throw new ServiceException("群员姓名不能为空");
                 GtdGroupMemberEntity groupMember = groupMemberRepository.findMemberByGroupIdAndUserId(groupId, gmDto.getUserId());
-                //GtdGroupMemberEntity groupMember =new GtdGroupMemberEntity();
-                groupMember.setUserId(gmDto.getUserId());
-                groupMember.setUserName(gmDto.getUserName());
-                groupMember.setUserContact(gmDto.getUserContact());
-                groupMember.setCreateId(userId);
-                groupMember.setUpdateDate(new Timestamp(new Date().getTime()));
-                groupMemberRepository.save(groupMember);
+                if(groupMember!=null) {
+                    groupMember.setUserId(gmDto.getUserId());
+                    groupMember.setUserName(gmDto.getUserName());
+                    groupMember.setUserContact(gmDto.getUserContact());
+                    groupMember.setCreateId(userId);
+                    groupMember.setUpdateDate(new Timestamp(new Date().getTime()));
+                    groupMemberRepository.save(groupMember);
+                }
             }
         } else {
             throw new ServiceException("权限群不可编辑");
@@ -774,6 +817,9 @@ public class GroupServicelmpl implements IGroupService {
         if (groupId == 0 || "".equals(groupId)) throw new ServiceException("群组ID不能为空");
         if (member.size() == 0 || member == null) throw new ServiceException("群员不能为空");
         GtdGroupEntity group = groupJpaRepository.findGtdGroupEntityByGroupId(groupId);
+        if (group == null) {
+            throw new ServiceException("该群组数据为空");
+        }
         boolean flag = false; //判断是否为权限群组
         Set<GtdLabelEntity> labels = group.getLabel();
         for (GtdLabelEntity label : labels) {
@@ -789,11 +835,13 @@ public class GroupServicelmpl implements IGroupService {
                 GtdGroupMemberEntity groupMember = groupMemberRepository.findMemberByGroupIdAndUserId(groupId, gmDto.getUserId());
                 //GtdGroupMemberEntity groupMember =new GtdGroupMemberEntity();
                 //修改群成员状态
-                //TODO 获取群成员状态
-                groupMember.setGroupMemberStatus(3);
-                groupMember.setUpdateId(userId);
-                groupMember.setUpdateDate(new Timestamp(new Date().getTime()));
-                groupMemberRepository.save(groupMember);
+                if(groupMember!=null) {
+                    //TODO 获取群成员状态
+                    groupMember.setGroupMemberStatus(3);
+                    groupMember.setUpdateId(userId);
+                    groupMember.setUpdateDate(new Timestamp(new Date().getTime()));
+                    groupMemberRepository.save(groupMember);
+                }
             }
         } else {
             throw new ServiceException("本地群成员无法修改");
